Move ud.ts to upper directory
ud.ts は react、angular 両方から利用されるため
